+++ frontend/src/main/resources/webapp/src/app/shared/app.menu.config.ts
@@ -1,7 +1,7 @@
 import {MenuRootItem} from 'ontimize-web-ngx';
 
 export const MENU_CONFIG: MenuRootItem[] = [
-  {id: 'home', name: 'HOME', icon: 'home', route: '/main/home'},
+  // {id: 'home', name: 'HOME', icon: 'home', route: '/main/home'},
   {id: 'books', name: 'BOOKS', icon: 'book', route: '/main/books'},
   {id: 'authors', name: 'AUTHORS', icon: 'people', route: '/main/authors'},
   {id: 'genres', name: 'GENRES', icon: 'style', route: '/main/genres'},