+++ frontend/src/main/resources/webapp/src/app/shared/services/login.guard.ts
@@ -1,0 +1,23 @@
+import {Injectable} from '@angular/core';
+import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
+import {Observable} from 'rxjs';
+import {LoginService} from "ontimize-web-ngx";
+
+@Injectable({
+  providedIn: 'root'
+})
+export class LoginGuard implements CanActivate {
+
+  constructor(public loginService: LoginService,
+              public router: Router) {}
+
+  canActivate(
+    next: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
+    if (!this.loginService.isLoggedIn()) {
+      this.router.navigate(['login']);
+      return false;
+    }
+    return true;
+  }
+}